Fix small errors in tests.

+++ ha/functional-java/src/test/java/ru/spbau/mit/bibaev/functional/func/Function1Test.java
@@ -11,7 +11,7 @@ public class Function1Test {
         Function1<Integer, Integer> square = value -> value * value;
         Function1<Integer, Integer> prev = value -> value - 1;
 
-        assertEquals(3, (long)square.compose(prev).apply(2));
-        assertEquals(1, (long)prev.compose(square).apply(2));
+        assertEquals(3, (int)square.compose(prev).apply(2));
+        assertEquals(1, (int)prev.compose(square).apply(2));
     }
 }
